move frontend view out of the api app
