+++ api/tacticalrmm/checks/views.py
@@ -4,7 +4,6 @@ from django.utils import timezone as djangotime
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework import status
-from rest_framework.authentication import TokenAuthentication
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.decorators import (
     api_view,
