+++ api/tacticalrmm/tacticalrmm/settings.py
@@ -11,7 +11,7 @@ AUTH_USER_MODEL = "accounts.User"
 
 # bump this version everytime vue code is changed
 # to alert user they need to manually refresh their browser
-APP_VER = "0.0.57"
+APP_VER = "0.0.58"
 
 # https://github.com/wh1te909/salt
 LATEST_SALT_VER = "1.1.0"
